+++ sumologic/resource_sumologic_connection_test.go
@@ -16,6 +16,7 @@ func TestAccConnection_create(t *testing.T) {
 	description := acctest.RandomWithPrefix("tf-connection-test-description")
 	url := "https://example.com"
 	defaultPayload := "{\"eventType\" : \"{{Name}}\"}"
+	resolutionPayload := "{\"eventType\" : \"{{Name}}\"}"
 	webhookType := "Webhook"
 
 	var connection Connection
@@ -26,7 +27,7 @@ func TestAccConnection_create(t *testing.T) {
 		CheckDestroy: testAccCheckConnectionDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: createConnectionConfig(name, connectionType, description, url, webhookType, defaultPayload),
+				Config: createConnectionConfig(name, connectionType, description, url, webhookType, defaultPayload, resolutionPayload),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckConnectionExists("sumologic_connection.test", &connection, t),
 					testAccCheckConnectionAttributes("sumologic_connection.test"),
@@ -35,6 +36,7 @@ func TestAccConnection_create(t *testing.T) {
 					resource.TestCheckResourceAttr("sumologic_connection.test", "description", description),
 					resource.TestCheckResourceAttr("sumologic_connection.test", "url", url),
 					resource.TestCheckResourceAttr("sumologic_connection.test", "default_payload", defaultPayload+"\n"),
+					resource.TestCheckResourceAttr("sumologic_connection.test", "resolution_payload", resolutionPayload+"\n"),
 					resource.TestCheckResourceAttr("sumologic_connection.test", "webhook_type", webhookType),
 				),
 			},
@@ -48,10 +50,49 @@ func TestAccConnection_createServiceNowWebhook(t *testing.T) {
 	description := acctest.RandomWithPrefix("tf-servicenow-webhook-connection-test-description")
 	url := "https://example.com"
 	defaultPayload := "{\"eventType\" : \"{{Name}}\"}"
+	resolutionPayload := "{\n  \"sys_id\": \"{{SysId}}\",\n  \"state\": \"6\",\n  \"incident_state\": \"6\",\n  \"work_notes\": \"Resolved value - Test\"\n}"
 	webhookType := "ServiceNow"
 	connectionSubtype := "Incident"
 
 	var connection Connection
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckConnectionDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: createServiceNowWebhookConnectionConfigWithResolutionPayload(name, connectionType, description, url,
+					connectionSubtype, defaultPayload, resolutionPayload),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckConnectionExists("sumologic_connection.serviceNowTest", &connection, t),
+					testAccCheckConnectionAttributes("sumologic_connection.serviceNowTest"),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "type", connectionType),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "name", name),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "description", description),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "url", url),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "default_payload", defaultPayload+"\n"),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "resolution_payload", resolutionPayload+"\n"),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "webhook_type", webhookType),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "connection_subtype", connectionSubtype),
+				),
+			},
+		},
+	})
+}
+
+func TestAccConnection_update(t *testing.T) {
+	var connection Connection
+	connectionType := "WebhookConnection"
+	name := acctest.RandomWithPrefix("tf-servicenow-webhook-connection-test-name")
+	description := acctest.RandomWithPrefix("tf-connection-test-description")
+	url := "https://example.com"
+	defaultPayload := "{\"eventType\" : \"{{Name}}\"}"
+	webhookType := "ServiceNow"
+	connectionSubtype := "Incident"
+
+	updatedDefaultPayload := `{"eventType" : "{{Name}} default updated"}`
+	updatedResolutionPayload := "{\n  \"sys_id\": \"{{SysId}}\",\n  \"state\": \"6\",\n  \"incident_state\": \"6\",\n  \"work_notes\": \"Resolved value - Test\"\n}"
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
@@ -71,49 +112,20 @@ func TestAccConnection_createServiceNowWebhook(t *testing.T) {
 					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "webhook_type", webhookType),
 					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "connection_subtype", connectionSubtype),
 				),
-			},
-		},
-	})
-}
-
-func TestAccConnection_update(t *testing.T) {
-	var connection Connection
-	connectionType := "WebhookConnection"
-	name := acctest.RandomWithPrefix("tf-connection-test-name")
-	url := "https://example.com"
-	defaultPayload := `{"eventType" : "{{Name}}"}`
-	webhookType := "Webhook"
-	fDescription := acctest.RandomWithPrefix("tf-connection-test-description")
-	sDescription := acctest.RandomWithPrefix("tf-connection-test-description")
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckConnectionDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: createConnectionConfig(name, connectionType, fDescription, url, webhookType, defaultPayload),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckConnectionExists("sumologic_connection.test", &connection, t),
-					testAccCheckConnectionAttributes("sumologic_connection.test"),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "type", connectionType),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "name", name),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "description", fDescription),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "url", url),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "default_payload", defaultPayload+"\n"),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "webhook_type", webhookType),
-				),
 			}, {
-				Config: createConnectionConfig(name, connectionType, sDescription, url, webhookType, defaultPayload),
+				Config: createServiceNowWebhookConnectionConfigWithResolutionPayload(name, connectionType, description, url, connectionSubtype,
+					updatedDefaultPayload, updatedResolutionPayload),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckConnectionExists("sumologic_connection.test", &connection, t),
-					testAccCheckConnectionAttributes("sumologic_connection.test"),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "type", connectionType),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "name", name),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "description", sDescription),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "url", url),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "default_payload", defaultPayload+"\n"),
-					resource.TestCheckResourceAttr("sumologic_connection.test", "webhook_type", webhookType),
+					testAccCheckConnectionExists("sumologic_connection.serviceNowTest", &connection, t),
+					testAccCheckConnectionAttributes("sumologic_connection.serviceNowTest"),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "type", connectionType),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "name", name),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "description", description),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "url", url),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "default_payload", updatedDefaultPayload+"\n"),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "resolution_payload", updatedResolutionPayload+"\n"),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "webhook_type", webhookType),
+					resource.TestCheckResourceAttr("sumologic_connection.serviceNowTest", "connection_subtype", connectionSubtype),
 				),
 			},
 		},
@@ -171,7 +183,7 @@ func testAccCheckConnectionDestroy(s *terraform.State) error {
 	return nil
 }
 
-func createConnectionConfig(name, connectionType, desc, url, webhookType, defaultPayload string) string {
+func createConnectionConfig(name, connectionType, desc, url, webhookType, defaultPayload string, resolutionPayload string) string {
 	return fmt.Sprintf(`
 resource "sumologic_connection" "test" {
 	name = "%s"
@@ -182,8 +194,11 @@ resource "sumologic_connection" "test" {
 	default_payload = <<JSON
 %s
 JSON
+	resolution_payload = <<JSON
+%s
+JSON
 }
-`, name, connectionType, desc, url, webhookType, defaultPayload)
+`, name, connectionType, desc, url, webhookType, defaultPayload, resolutionPayload)
 }
 
 func createServiceNowWebhookConnectionConfig(name, connectionType, desc, url, connectionSubtype, defaultPayload string) string {
@@ -203,4 +218,25 @@ resource "sumologic_connection" "serviceNowTest" {
 JSON
 }
 `, name, connectionType, desc, url, connectionSubtype, defaultPayload)
+}
+func createServiceNowWebhookConnectionConfigWithResolutionPayload(name, connectionType, desc, url, connectionSubtype, defaultPayload string, resolutionPayload string) string {
+	return fmt.Sprintf(`
+resource "sumologic_connection" "serviceNowTest" {
+	name = "%s"
+	type = "%s"
+	description = "%s"
+	url = "%s"
+	headers = {
+		"Authorization": "Basic SOMERANDOMAUTHSTRING"
+	}
+	webhook_type = "ServiceNow"
+	connection_subtype = "%s"
+	default_payload = <<JSON
+%s
+JSON
+	resolution_payload = <<JSON
+%s
+JSON
+}
+`, name, connectionType, desc, url, connectionSubtype, defaultPayload, resolutionPayload)
 }