+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 2.21.1 (Unreleased)
+FEATURES:
+* Add new optional `resolution_payload` field to connection resource and `resolution_payload_override` field to notifications section of monitor resource (GH-482)
 
 ## 2.21.0 (February 27, 2023)
 FEATURES: